server/requirements-dev.txt
-r requirements.txt

alembic==1.16.5
Flask-Migrate==4.1.0
Mako==1.3.10
pytest==8.4.2
pytest-cov==6.2.1
